Renamed morgan.ts to requestLogging.ts
